main components splited
