+++ client/src/components/offers/list.item.styles.scss
@@ -1,6 +1,6 @@
 .offer-item {
-  width: 25rem;
-  height: 79vh;
+  width: 30%;
+  //height: 79vh;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -42,7 +42,7 @@
 
   .img-container {
     width: 100%;
-    height: 20rem;
+    height: 30rem;
     margin-bottom: 1.2rem;
     
     img {
