+++ client/src/components/offers/offers.component.styles.scss
@@ -1,14 +1,15 @@
 .offers-wrapper {
-  height: 95vh;
-  width: 100%;
+  grid-column: 1 / 10;
+  grid-row: 1;
+  border-right: 1.5px dashed black;
   border-top-left-radius: 30px;
   border-bottom-left-radius: 30px;
   background-color: #fcfcfc;
+
   display: flex;
   flex-direction: column;
   align-items: center;
   //border: 1px solid teal;
-  padding: 1rem;
 
   h2 {
     margin-bottom: 1rem;
@@ -32,6 +33,7 @@
     display: flex;
     width: 98%;
   }
+
   .close-circle-outline ion-icon {
     margin-left: auto;
     margin-bottom: 1.5rem;
@@ -44,10 +46,10 @@
     }
   }
 }
+
 .display {
   border-radius: 30px;
   padding: 1rem;
-
   //background-color: #f5f5f5;
   width: 97%;
   display: flex;
@@ -55,6 +57,7 @@
   flex-direction: column;
   //border: 1px solid red;
 }
+
 .dddd {
   display: flex;
   align-items: center;
