+++ client/src/pages/authentication/authentication.jsx
@@ -1,18 +1,16 @@
-import React from 'react'
-
-import './authentication.styles.scss'
+import React from "react";
+import "./authentication.styles.scss";
 
 //signIn Component
-import SignIn from "../../components/signin/signin.component"
-import SignUp from "../../components/signup/signup.component"
+import SignIn from "../../components/signin/signin.component";
+import SignUp from "../../components/signup/signup.component";
 
-function Authentication() {
-    return (
-        <div className="authentication">
-            <SignIn />
-            <SignUp/>
-        </div>
-    )
-}
-
-export default Authentication
+const Authentication = () => {
+  return (
+    <div className="app__authentication">
+      <SignIn />
+      <SignUp />
+    </div>
+  );
+};
+export default Authentication;
