+++ client/src/pages/authentication/authentication.styles.scss
@@ -1,17 +1,24 @@
-.authentication {
-  /* position */
-  display: flex;
-  height: 95vh;
-  justify-content: space-around;
-  padding-top: 20vh;
-  align-items: flex-start;
-  background-color: #fcfcfc;
-  background-size: cover;
-  background-position: center;
-  color: #181818;
-  font-weight: 800;
+.app {
+  &__authentication {
+    /* grid columns and row*/
+    grid-column: 1 / 10;
+    grid-row: 1;
+    border-right: 1.5px dashed black;
+    //border: 1px solid blue;
 
-  // border-radius: 30px;
-  border-top-left-radius: 30px;
-  border-bottom-left-radius: 30px;
+    /* position */
+    display: flex;
+    justify-content: space-around;
+    padding-top: 20vh;
+    align-items: flex-start;
+    background-color: #fcfcfc;
+    background-size: cover;
+    background-position: center;
+    color: #181818;
+    font-weight: 800;
+
+    // border-radius: 30px;
+    border-top-left-radius: 30px;
+    border-bottom-left-radius: 30px;
+  }
 }
