+++ client/src/pages/home/main.styles.scss
@@ -1,0 +1,146 @@
+.app {
+  display: grid;
+  grid-template-rows: 96vh;
+  grid-template-columns: repeat(12, minmax(min-content, 16rem));
+
+  &__main {
+    grid-column: 1 / 10;
+    grid-row: 1;
+    //border: 1px solid blue;
+    display: grid;
+    grid-template-rows: repeat(12, minmax(min-content, 1fr));
+    grid-template-columns: repeat(12, minmax(min-content, 16rem));
+    background-color: #303030;
+    border-top-left-radius: 30px;
+    border-bottom-left-radius: 30px;
+
+    img {
+      z-index: -2;
+      grid-column: 1 / 10;
+      width: 100%;
+      height: auto;
+    }
+    //@debug: 30px;
+
+    &__logo {
+      grid-column: 2 / 7;
+      grid-row: 1 / 2;
+      //border: 1px solid pink;
+      font-size: 1.6vw;
+      letter-spacing: 0.8px;
+      display: flex;
+      align-items: center;
+
+      span {
+        color: #f75d37;
+      }
+    }
+
+    &__intro {
+      grid-column: 2 / 5;
+      grid-row: 5 / 8;
+      //border: 1px solid blue;
+      //align-self: center;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-evenly;
+
+      h1 {
+      }
+
+      h5 {
+        font-size: 1.4rem;
+        font-weight: 600;
+      }
+    }
+
+    a {
+      //border: 1px solid brown;
+      grid-column: 2 / 3;
+      grid-row: 8 / 9;
+
+      .btn {
+        display: inline-block;
+        font-weight: 400;
+        color: #212529;
+        text-align: center;
+        vertical-align: middle;
+        user-select: none;
+        background-color: transparent;
+        border: 1px solid transparent;
+        padding: 0.375rem 0.75rem;
+        font-size: 1.6rem;
+        line-height: 1.5;
+        border-radius: 0.25rem;
+        transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out,
+          border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+      }
+
+      @media (prefers-reduced-motion: reduce) {
+        .btn {
+          transition: none;
+        }
+      }
+
+      .btn:hover {
+        color: #212529;
+        text-decoration: none;
+      }
+
+      .btn:focus,
+      .btn.focus {
+        outline: 0;
+        box-shadow: 0 0 0 0.2rem rgba(0, 123, 255, 0.25);
+      }
+
+      .btn.disabled,
+      .btn:disabled {
+        opacity: 0.65;
+      }
+
+      .btn:not(:disabled):not(.disabled) {
+        cursor: pointer;
+      }
+
+      a.btn.disabled,
+      fieldset:disabled a.btn {
+        pointer-events: none;
+      }
+
+      .btn-primary {
+        color: #fafafa;
+        background-color: #f75d37;
+        border-color: #f75d37;
+        opacity: 0.9;
+        width: 10vw;
+      }
+
+      .btn-primary:hover {
+        color: #fafafa;
+        opacity: 1;
+      }
+
+      .btn-primary:focus,
+      .btn-primary.focus {
+        color: #fafafa;
+        background-color: #0069d9;
+        border-color: #0062cc;
+        box-shadow: 0 0 0 0.2rem rgba(38, 143, 255, 0.5);
+      }
+
+      .btn-primary:not(:disabled):not(.disabled):active:focus,
+      .btn-primary:not(:disabled):not(.disabled).active:focus,
+      .show > .btn-primary.dropdown-toggle:focus {
+        box-shadow: 0 0 0 0.2rem rgba(38, 143, 255, 0.5);
+      }
+
+      .btn-lg,
+      .btn-group-lg > .btn {
+        padding: 0.5rem 1rem;
+        font-size: 1.8rem;
+        line-height: 1.5;
+        border-radius: 0.3rem;
+      }
+    }
+  }
+}
